requirements.txt
Flask==2.1.2
Werkzeug==2.1.1
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.0.4
Pillow==9.0.1
torch==1.11.0
transformers==4.19.2
requests==2.27.1
gunicorn==20.1.0